Trying to set the odometry at 100Hz... Not working. Problem with angle odometry, what unit ?

+++ mainboard/firmware/odometry.h
@@ -11,9 +11,9 @@
 
 
 struct position {
-  float xpos;
-  float ypos;
-  float ang;
+  double xpos;
+  double ypos;
+  double ang;
 };
